refactor: move remote read Firestore access to entities
