+++ src/entities/deck/api/firestore.ts
@@ -1,13 +1,63 @@
-import type { Deck, DeckCreate, DeckCreateInput, DeckEdit, DeleteDeckInput, EditDeckInput } from "../model/types";
+import type {
+  Deck,
+  DeckCreate,
+  DeckCreateInput,
+  DeckEdit,
+  DeckId,
+  DeleteDeckInput,
+  EditDeckInput,
+} from "../model/types";
 
-import { collection, deleteDoc, doc, getDocs, query, setDoc, updateDoc, where } from "firebase/firestore";
+import { collection, deleteDoc, doc, getDocs, onSnapshot, query, setDoc, updateDoc, where } from "firebase/firestore";
+import { z } from "zod";
 
 import { db } from "@/shared/firebase";
-import { getTimestamp, omitUndefined } from "@/shared/firestore";
+import { getTimestamp, omitUndefined, parseFirestoreDocument } from "@/shared/firestore";
 import { createDeckSchema, deleteDeckSchema, editDeckSchema } from "../model/schema";
+import { replaceDecks } from "../model/store";
 
 const DECK_COLLECTION = "deck";
 const CARD_COLLECTION = "card";
+
+const deckDtoSchema = z.object({
+  id: z.string().optional(),
+  name: z.string(),
+  url: z.string().optional(),
+  isPublic: z.boolean(),
+  uid: z.string(),
+  createdAt: z.number(),
+  updatedAt: z.number(),
+  deletedAt: z.number().nullable(),
+  scoreMax: z.number().nullable(),
+  scoreMin: z.number().nullable(),
+  selectedTags: z.array(z.string()),
+  tagAndFilter: z.boolean(),
+  category: z.string(),
+  convertToBr: z.boolean(),
+});
+
+const convertDeckDtoToDeck = (id: DeckId, value: unknown): Deck => {
+  const dto = parseFirestoreDocument(deckDtoSchema, "deck", id, value);
+  const deck: Deck = { ...dto, id };
+  if (dto.url === undefined) delete deck.url;
+  return deck;
+};
+
+export const subscribeDecks = (uid: string, onError: (error: Error) => void): (() => void) =>
+  onSnapshot(
+    query(collection(db, DECK_COLLECTION), where("uid", "==", uid)),
+    (snapshot) => {
+      try {
+        const decks = snapshot.docs
+          .map((document) => convertDeckDtoToDeck(document.id, document.data()))
+          .filter((deck) => deck.deletedAt === null);
+        replaceDecks(decks);
+      } catch (cause) {
+        onError(cause instanceof Error ? cause : new Error(String(cause)));
+      }
+    },
+    onError
+  );
 
 export const generateDeckId = (): string => doc(collection(db, DECK_COLLECTION)).id;
