+++ src/entities/deck/api/subscription.spec.tsx
@@ -1,17 +1,28 @@
 import { act, renderHook } from "@testing-library/react";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
-import { clearDecks, useDecks } from "@/entities/deck";
+import { useDecks } from "../model/hooks";
+import { clearDecks } from "../model/store";
+
+type TestDocument = { id: string; data: () => Record<string, unknown> };
+type TestSnapshot = { docs: TestDocument[] };
 
 const mocks = vi.hoisted(() => ({
   collection: vi.fn((...parts: unknown[]) => parts),
-  onSnapshot: vi.fn(),
   query: vi.fn((...parts: unknown[]) => parts),
   where: vi.fn((...parts: unknown[]) => parts),
+  next: undefined as ((snapshot: TestSnapshot) => void) | undefined,
+  error: undefined as ((error: Error) => void) | undefined,
   unsubscribe: vi.fn(),
+  onSnapshot: vi.fn((_query: unknown, next: (snapshot: TestSnapshot) => void, error: (cause: Error) => void) => {
+    mocks.next = next;
+    mocks.error = error;
+    return mocks.unsubscribe;
+  }),
 }));
 
-vi.mock("firebase/firestore", () => ({
+vi.mock("firebase/firestore", async (importOriginal) => ({
+  ...(await importOriginal<typeof import("firebase/firestore")>()),
   collection: mocks.collection,
   onSnapshot: mocks.onSnapshot,
   query: mocks.query,
@@ -19,9 +30,9 @@ vi.mock("firebase/firestore", () => ({
 }));
 vi.mock("@/shared/firebase", () => ({ db: "db" }));
 
-import { subscribeDecks } from "./deck";
+import { subscribeDecks } from "./firestore";
 
-const deckDocument = (id: string, overrides: Record<string, unknown> = {}) => ({
+const deckDocument = (id: string, overrides: Record<string, unknown> = {}): TestDocument => ({
   id,
   data: () => ({
     name: "Remote Deck",
@@ -40,30 +51,31 @@ const deckDocument = (id: string, overrides: Record<string, unknown> = {}) => ({
   }),
 });
 
-const getSnapshotHandler = () =>
-  mocks.onSnapshot.mock.calls[0]?.[1] as (snapshot: { docs: ReturnType<typeof deckDocument>[] }) => void;
-const getErrorHandler = () => mocks.onSnapshot.mock.calls[0]?.[2] as (error: Error) => void;
-
-describe("Deck app synchronization", () => {
+describe("Deck Firestore subscription", () => {
   beforeEach(() => {
     clearDecks();
     vi.clearAllMocks();
-    mocks.onSnapshot.mockReturnValue(mocks.unsubscribe);
+    mocks.next = undefined;
+    mocks.error = undefined;
   });
 
-  it("subscribes by UID and replaces the store with active Decks", () => {
+  it("subscribes by UID and fully replaces active Decks from each snapshot", () => {
     const { result } = renderHook(useDecks);
     const unsubscribe = subscribeDecks("uid-a", vi.fn());
 
     expect(mocks.collection).toHaveBeenCalledWith("db", "deck");
     expect(mocks.where).toHaveBeenCalledWith("uid", "==", "uid-a");
     act(() =>
-      getSnapshotHandler()({
+      mocks.next?.({
         docs: [deckDocument("active", { url: "https://example.com" }), deckDocument("deleted", { deletedAt: 3 })],
       })
     );
 
     expect(result.current).toEqual([expect.objectContaining({ id: "active", url: "https://example.com" })]);
+
+    act(() => mocks.next?.({ docs: [deckDocument("replacement", { name: "Current" })] }));
+    expect(result.current).toEqual([expect.objectContaining({ id: "replacement", name: "Current" })]);
+
     unsubscribe();
     expect(mocks.unsubscribe).toHaveBeenCalledOnce();
   });
@@ -72,7 +84,7 @@ describe("Deck app synchronization", () => {
     const onError = vi.fn();
     subscribeDecks("uid-a", onError);
 
-    act(() => getSnapshotHandler()({ docs: [deckDocument("invalid", { selectedTags: [42] })] }));
+    act(() => mocks.next?.({ docs: [deckDocument("invalid", { selectedTags: [42] })] }));
 
     expect(onError).toHaveBeenCalledWith(
       expect.objectContaining({ name: "FirestoreDocumentValidationError", documentId: "invalid" })
@@ -81,10 +93,10 @@ describe("Deck app synchronization", () => {
 
   it("reports Firestore subscription errors", () => {
     const onError = vi.fn();
-    subscribeDecks("uid-a", onError);
     const error = new Error("subscription failed");
+    subscribeDecks("uid-a", onError);
 
-    getErrorHandler()(error);
+    mocks.error?.(error);
 
     expect(onError).toHaveBeenCalledWith(error);
   });
