+++ src/entities/deck/index.ts
@@ -1,5 +1,6 @@
-export { createDeck, deleteDeck, editDeck, generateDeckId } from "./api/firestore";
+export { createDeck, deleteDeck, editDeck, generateDeckId, subscribeDecks } from "./api/firestore";
 export { CATEGORY, getCategory, isHighlightLanguage } from "./model/rules";
 export { useDeck, useDecks } from "./model/hooks";
+/** @public */
 export { clearDecks, replaceDecks } from "./model/store";
 export type { Deck, DeckCreateInput, DeckEdit, DeckId } from "./model/types";
